Split event receivers and their unit tests into different files
